Move GUI tests into tests/ dir and fix conftest related stuff so everything loads what it needs and passes
