+++ tests/local_onionshare_share_mode_slug_persistent_test.py
