+++ tests/onionshare_share_mode_download_stay_open_test.py
