+++ tests/onionshare_share_mode_persistent_test.py
