+++ tests/onionshare_share_mode_stealth_test.py
