+++ tests/onionshare_share_mode_tor_connection_killed_test.py
